+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `@reevit/typescript` will be documented in this file.
 
+## [Unreleased] - 2026-02-04
+
+### 🛠 Improvements
+
+- Added `idempotencyKey` support for payment intent creation.
+
 ## [0.5.0] - 2026-01-11
 
 ### 🚀 New Features